# services/api/openapi/openapi.v1.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Decider API",
    "version": "1.0.0"
  },
  "paths": {
    "/api/v1/health": {
      "get": {
        "tags": [
          "v1"
        ],
        "summary": "Get Health V1",
        "operationId": "get_health_v1_api_v1_health_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HealthResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/context": {
      "get": {
        "tags": [
          "v1"
        ],
        "summary": "Get Auth Context V1",
        "operationId": "get_auth_context_v1_api_v1_auth_context_get",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthContextResponse"
                }
              }
            }
          },
          "401": {
            "description": "Invalid or expired token."
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/tenants/{tenant_id}/resources": {
      "get": {
        "tags": [
          "v1"
        ],
        "summary": "Get Tenant Resources V1",
        "operationId": "get_tenant_resources_v1_api_v1_tenants__tenant_id__resources_get",
        "parameters": [
          {
            "name": "tenant_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Tenant Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TenantResourcesResponse"
                }
              }
            }
          },
          "401": {
            "description": "Invalid or expired token."
          },
          "403": {
            "description": "Forbidden"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/tenants/{tenant_id}/exports": {
      "post": {
        "tags": [
          "v1"
        ],
        "summary": "Post Tenant Export V1",
        "operationId": "post_tenant_export_v1_api_v1_tenants__tenant_id__exports_post",
        "parameters": [
          {
            "name": "tenant_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Tenant Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExportResponse"
                }
              }
            }
          },
          "401": {
            "description": "Invalid or expired token."
          },
          "403": {
            "description": "Forbidden"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/tenants/{tenant_id}/entitlements/{subject}": {
      "get": {
        "tags": [
          "v1"
        ],
        "summary": "Get Tenant Entitlements V1",
        "operationId": "get_tenant_entitlements_v1_api_v1_tenants__tenant_id__entitlements__subject__get",
        "parameters": [
          {
            "name": "tenant_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Tenant Id"
            }
          },
          {
            "name": "subject",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Subject"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EntitlementsResponse"
                }
              }
            }
          },
          "401": {
            "description": "Invalid or expired token."
          },
          "403": {
            "description": "Forbidden"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "v1"
        ],
        "summary": "Put Tenant Entitlements V1",
        "operationId": "put_tenant_entitlements_v1_api_v1_tenants__tenant_id__entitlements__subject__put",
        "parameters": [
          {
            "name": "tenant_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Tenant Id"
            }
          },
          {
            "name": "subject",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Subject"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EntitlementsUpdateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EntitlementsResponse"
                }
              }
            }
          },
          "401": {
            "description": "Invalid or expired token."
          },
          "403": {
            "description": "Forbidden"
          },
          "422": {
            "description": "Validation error."
          }
        }
      }
    },
    "/api/v1/tenants/{tenant_id}/audit/events": {
      "get": {
        "tags": [
          "v1"
        ],
        "summary": "Get Tenant Audit Events V1",
        "operationId": "get_tenant_audit_events_v1_api_v1_tenants__tenant_id__audit_events_get",
        "parameters": [
          {
            "name": "tenant_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Tenant Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TenantAuditEventsResponse"
                }
              }
            }
          },
          "401": {
            "description": "Invalid or expired token."
          },
          "403": {
            "description": "Forbidden"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "AuditEventResponse": {
        "properties": {
          "event_id": {
            "type": "string",
            "title": "Event Id"
          },
          "action": {
            "type": "string",
            "enum": [
              "entitlements.updated",
              "export.requested"
            ],
            "title": "Action"
          },
          "actor_subject": {
            "type": "string",
            "title": "Actor Subject"
          },
          "target_subject": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target Subject"
          },
          "tenant_id": {
            "type": "string",
            "title": "Tenant Id"
          },
          "outcome": {
            "type": "string",
            "enum": [
              "success",
              "forbidden"
            ],
            "title": "Outcome"
          },
          "occurred_at": {
            "type": "string",
            "title": "Occurred At"
          },
          "reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          }
        },
        "type": "object",
        "required": [
          "event_id",
          "action",
          "actor_subject",
          "tenant_id",
          "outcome",
          "occurred_at"
        ],
        "title": "AuditEventResponse"
      },
      "AuthContextResponse": {
        "properties": {
          "authenticated": {
            "type": "boolean",
            "title": "Authenticated"
          },
          "subject": {
            "type": "string",
            "title": "Subject"
          },
          "tenant_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tenant Id"
          },
          "scopes": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Scopes"
          },
          "roles": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Roles"
          },
          "module_entitlements": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Module Entitlements"
          }
        },
        "type": "object",
        "required": [
          "authenticated",
          "subject",
          "tenant_id"
        ],
        "title": "AuthContextResponse"
      },
      "EntitlementAuditMetadata": {
        "properties": {
          "event_id": {
            "type": "string",
            "title": "Event Id"
          },
          "action": {
            "type": "string",
            "const": "entitlements.updated",
            "title": "Action"
          },
          "actor_subject": {
            "type": "string",
            "title": "Actor Subject"
          },
          "target_subject": {
            "type": "string",
            "title": "Target Subject"
          },
          "tenant_id": {
            "type": "string",
            "title": "Tenant Id"
          },
          "occurred_at": {
            "type": "string",
            "title": "Occurred At"
          }
        },
        "type": "object",
        "required": [
          "event_id",
          "action",
          "actor_subject",
          "target_subject",
          "tenant_id",
          "occurred_at"
        ],
        "title": "EntitlementAuditMetadata"
      },
      "EntitlementsResponse": {
        "properties": {
          "tenant_id": {
            "type": "string",
            "title": "Tenant Id"
          },
          "subject": {
            "type": "string",
            "title": "Subject"
          },
          "enabled_modules": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Enabled Modules"
          },
          "audit_metadata": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/EntitlementAuditMetadata"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "tenant_id",
          "subject"
        ],
        "title": "EntitlementsResponse"
      },
      "EntitlementsUpdateRequest": {
        "properties": {
          "enabled_modules": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Enabled Modules"
          }
        },
        "type": "object",
        "title": "EntitlementsUpdateRequest"
      },
      "ExportAuditMetadata": {
        "properties": {
          "event_id": {
            "type": "string",
            "title": "Event Id"
          },
          "action": {
            "type": "string",
            "const": "export.requested",
            "title": "Action"
          },
          "actor_subject": {
            "type": "string",
            "title": "Actor Subject"
          },
          "tenant_id": {
            "type": "string",
            "title": "Tenant Id"
          },
          "outcome": {
            "type": "string",
            "enum": [
              "success",
              "forbidden"
            ],
            "title": "Outcome"
          },
          "occurred_at": {
            "type": "string",
            "title": "Occurred At"
          },
          "reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          }
        },
        "type": "object",
        "required": [
          "event_id",
          "action",
          "actor_subject",
          "tenant_id",
          "outcome",
          "occurred_at"
        ],
        "title": "ExportAuditMetadata"
      },
      "ExportResponse": {
        "properties": {
          "tenant_id": {
            "type": "string",
            "title": "Tenant Id"
          },
          "export_id": {
            "type": "string",
            "title": "Export Id"
          },
          "status": {
            "type": "string",
            "const": "accepted",
            "title": "Status"
          },
          "audit_metadata": {
            "$ref": "#/components/schemas/ExportAuditMetadata"
          }
        },
        "type": "object",
        "required": [
          "tenant_id",
          "export_id",
          "status",
          "audit_metadata"
        ],
        "title": "ExportResponse"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "HealthResponse": {
        "properties": {
          "status": {
            "type": "string",
            "const": "ok",
            "title": "Status"
          }
        },
        "type": "object",
        "required": [
          "status"
        ],
        "title": "HealthResponse"
      },
      "TenantAuditEventsResponse": {
        "properties": {
          "tenant_id": {
            "type": "string",
            "title": "Tenant Id"
          },
          "events": {
            "items": {
              "$ref": "#/components/schemas/AuditEventResponse"
            },
            "type": "array",
            "title": "Events"
          }
        },
        "type": "object",
        "required": [
          "tenant_id"
        ],
        "title": "TenantAuditEventsResponse"
      },
      "TenantBaseResource": {
        "properties": {
          "resource_id": {
            "type": "string",
            "title": "Resource Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          }
        },
        "type": "object",
        "required": [
          "resource_id",
          "name"
        ],
        "title": "TenantBaseResource"
      },
      "TenantResourcesResponse": {
        "properties": {
          "tenant_id": {
            "type": "string",
            "title": "Tenant Id"
          },
          "resources": {
            "items": {
              "$ref": "#/components/schemas/TenantBaseResource"
            },
            "type": "array",
            "title": "Resources"
          }
        },
        "type": "object",
        "required": [
          "tenant_id"
        ],
        "title": "TenantResourcesResponse"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}